+++ packages/polaris-schema/src/directives/permissions-type-defs.ts
@@ -1,5 +1,5 @@
 import gql from 'graphql-tag';
 
-export const directivesTypeDefs = gql`
+export const permissionsTypeDefs = gql`
     directive @permissions(entityTypes: [String], actions: [String]) on FIELD | FIELD_DEFINITION
 `;