+++ packages/polaris-core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.63](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-core@2.0.0-beta.62...@enigmatis/polaris-core@2.0.0-beta.63) (2021-03-07)
+
+
+### Features
+
+* Configurable polaris schema ([#65](https://github.com/Enigmatis/polaris-united/issues/65)) ([fa5afbd](https://github.com/Enigmatis/polaris-united/commit/fa5afbdec5e583d9d2a98d39b3bdede07bc94524))
+
+
+
+
+
+# [2.0.0-beta.62](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-core@2.0.0-beta.61...@enigmatis/polaris-core@2.0.0-beta.62) (2021-03-04)
+
+
+### Bug Fixes
+
+* transaction to close on auto snapshot ([#68](https://github.com/Enigmatis/polaris-united/issues/68)) ([4c7519b](https://github.com/Enigmatis/polaris-united/commit/4c7519b8e10f8bc972b206bfc50a932f4153d7f7))
+
+
+
+
+
 # [2.0.0-beta.61](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-core@2.0.0-beta.60...@enigmatis/polaris-core@2.0.0-beta.61) (2021-02-23)