+++ packages/polaris-nest/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-nest@1.7.2...@enigmatis/polaris-nest@1.8.0) (2021-03-07)
+
+
+### Features
+
+* Configurable polaris schema ([#65](https://github.com/Enigmatis/polaris-united/issues/65)) ([fa5afbd](https://github.com/Enigmatis/polaris-united/commit/fa5afbdec5e583d9d2a98d39b3bdede07bc94524))
+
+
+
+
+
+## [1.7.2](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-nest@1.7.1...@enigmatis/polaris-nest@1.7.2) (2021-03-04)
+
+**Note:** Version bump only for package @enigmatis/polaris-nest
+
+
+
+
+
 ## [1.7.1](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-nest@1.7.0...@enigmatis/polaris-nest@1.7.1) (2021-02-23)
 
 **Note:** Version bump only for package @enigmatis/polaris-nest